#16 some code clean up
