--- app/src/main/java/ir/thatsmejavad/backgroundable/core/viewmodel/Inject.kt
@@ -1,15 +1,0 @@
-package ir.thatsmejavad.backgroundable.core.viewmodel
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-
-@Composable
-fun Inject(
-    viewModelFactory: ViewModelFactory,
-    content: @Composable () -> Unit
-) {
-    CompositionLocalProvider(
-        LocalViewModelFactory provides viewModelFactory,
-        content = content
-    )
-}
